AD-HOC fix (5.6): Add a dist upgrade to the root FS during build
Quay is showing some vulns that have not been fixed in the upstream
build, but that are "fixable". So, we add a dist-upgrade to the build
step, so whatever packages exist on the rootfs are upgraded to their
most recent versions.

Further changes were made to document the new versioning strategy