AD-HOC refactor (5.6): Initial work to reduce the maintenance of multiple images
Each of the images follows fundamentally the same pattern of
"installable PHP", varying only the PHP version.

The reason there are so many is that PHP follows an LTS model, and
applications are often slow to migrate between runtimes (famously
leading to many ancient applications on PHP)

This repo hopes to support all LTS that are currently within security
updates, but in an optimised and up to date way.

The goal of this work is to ship production ready, but slim images on
which others can be built upon.