AD-HOC refactor (Base): Create base image allowing easy version
The goal of this refactoring work more generally is to allow easy
updating of multiple, LTS PHP versions. Generally there will be several
versions of these, but they share much in the way of configuration; PHP
optimisations etc.

This commit separates out the provisioning script into its own base
image, which includes the optimisations etc. It allows installing
various versions of PHP, but still maintaining optimisations in a single
spot.

The intent is that this base image will be updated automatically by
rerunning the build each day, and subsequent images additionally
automatically updated.