AD-HOC refactor (*): Replace the multiple dockerflies with a single configurable
Currently there is a problem maintaining this work in that every time a
chance is made, it must be made in three places. This overhead is a
drag, and not super elegant to work with.

This commit replaces the three different trees with a single tree, and a
configurable build argument. This was only possible fairly recently,
with the introduction of the --build-arg from docker.

== Stakeholder Impact ==

=== Project Owner ====

This is a prerequisite to automating the builds of these images. Such
automation is in turn a requirement for publishing images without
security issues.

=== Developers ===

This should make it much simpler to maintain this repository over time.