AD-HOC feat (HTTPS): Add and enable HTTPS by default
HTTPS is now the defacto standard for transport of HTTP content, due to
safety and security concerns on the broader internet.

This commit introduces and configures HTTPS on the basic container, such
that implementing containers have "safe" defaults from which to draw
their own implementations.

== Design Notes ==

=== Modern Profile ===

The profile chosen to generate the HTTPS content is compatible with:

  - Firefox 27+
  - Chrome 30+
  - IE 11 on Windows 7+
  - Edge
  - Opera 17+
  - Safari 9+
  - Android 5.0+
  - Java 8+

The author does not wish to support services that would wish to connect
in browsers older than this.

=== Snake oil certificate ===

To make this work out of the box with all clients a valid certificate is
required. The author implemented the commonly used "snake oil"
certificate.

While this certificate is not issued by a valid CA (indeed, it is
generated during installation) it's useful to verify the HTTP/2
connection type.

=== Enabling HTTP/2 ===

While HTTP/2 was possible prior to this work, it was not enabled by
default. Now that the container packages HTTPS connections by default,
it makes sense to also enable HTTP/2 by default.

=== Enabling HSTS ===

This commit also includes changes to enable mod_header and HSTS.
Connections that do not include HSTS headers are vulnerable to tools
like SSL strip.

Note: This relies on the user having connected the site at least 1 time
over a safe connection that does not strip SSL.