AD-HOC feat (ssmtp): Add SSMTP to stack
An extremely common problem that must be solved in PHP is the
requirement to send mail. This commit adds the mta SSMTP to the
application, which will forward mail immediately in the invocation of
the sendmail binary. It thus works as an effective shim between the
single process of Apache and "real" MTAs